+++ packages/auto-merge/src/AutoMerge.test.ts
@@ -2,11 +2,12 @@ import {expect, describe, test, beforeEach, afterEach, vi} from 'vitest';
 import nock from 'nock';
 import {StatusCodes as HTTP_STATUS} from 'http-status-codes';
 
-import {AutoApprover, Repository} from '../src/AutoApprover.js';
+import {AutoMerge} from './AutoMerge.js';
+import type {GitHubPullRequest, Repository} from './types/index.js';
 
-describe('AutoApprover', () => {
+describe('AutoMerge', () => {
   describe('checkRepository', () => {
-    let autoApprover: AutoApprover;
+    let autoMerge: AutoMerge;
 
     const validRepositoryNames = ['ffflorian/example-project-1', 'my-name/my_great-project42342'];
     const repositories: Repository[] = [
@@ -21,7 +22,7 @@ describe('AutoApprover', () => {
             number: 253,
             title: 'chore: bump eslint-plugin-typescript-sort-keys from 1.3.0 to 1.5.0',
           },
-        ],
+        ] as GitHubPullRequest[],
         repositorySlug: validRepositoryNames[0],
       },
       {
@@ -35,27 +36,32 @@ describe('AutoApprover', () => {
             number: 252,
             title: 'chore: bump typescript from 4.0.2 to 4.0.3',
           },
-        ],
+        ] as GitHubPullRequest[],
         repositorySlug: validRepositoryNames[1],
       },
     ];
 
     beforeEach(() => {
-      autoApprover = new AutoApprover({
+      autoMerge = new AutoMerge({
         authToken: 'example-token',
         projects: {
           gitHub: validRepositoryNames,
         },
       });
 
-      nock(autoApprover['apiClient'].defaults.baseURL!)
-        .post(/^\/repos\/[^/]+\/[^/]+\/pulls\/\d+\/reviews\/?$/)
+      nock(autoMerge['apiClient'].defaults.baseURL!)
+        .post(/^\/repos\/.+?\/.+?\/pulls\/\d+\/reviews\/?$/)
         .reply(HTTP_STATUS.OK, {data: 'not-used'})
         .persist();
 
-      nock(autoApprover['apiClient'].defaults.baseURL!)
-        .post(/^\/repos\/[^/]+\/[^/]+\/issues\/\d+\/comments\/?$/)
-        .reply(HTTP_STATUS.CREATED, {data: 'not-used'})
+      nock(autoMerge['apiClient'].defaults.baseURL!)
+        .get(/^\/repos\/.+?\/.+?\/pulls\/\d+\/?$/)
+        .reply(HTTP_STATUS.OK, {data: 'not-used'})
+        .persist();
+
+      nock(autoMerge['apiClient'].defaults.baseURL!)
+        .put(/^\/repos\/.+?\/.+?\/pulls\/\d+\/merge\/?$/)
+        .reply(HTTP_STATUS.OK, {data: 'not-used'})
         .persist();
     });
 
@@ -65,7 +71,7 @@ describe('AutoApprover', () => {
 
     test('detects valid repository names', () => {
       validRepositoryNames.forEach(repositoryName => {
-        const checkResult = autoApprover['checkRepositorySlug'](repositoryName);
+        const checkResult = autoMerge['checkRepositorySlug'](repositoryName);
         expect(checkResult).toBe(true);
       });
     });
@@ -79,7 +85,7 @@ describe('AutoApprover', () => {
         'valid-username/invalid!project-name',
       ];
       invalidRepositoryNames.forEach(repositoryName => {
-        const checkResult = autoApprover['checkRepositorySlug'](repositoryName);
+        const checkResult = autoMerge['checkRepositorySlug'](repositoryName);
         expect(checkResult).toBe(false);
       });
     });
@@ -87,42 +93,41 @@ describe('AutoApprover', () => {
     describe('getMatchingRepositories', () => {
       test('matches repositories', () => {
         const expectedESLint = [repositories[0]];
-        const actualESLint = autoApprover['getMatchingRepositories'](repositories, /eslint/gi);
+        const actualESLint = autoMerge['getMatchingRepositories'](repositories, /eslint/gi);
         expect(actualESLint).toEqual(expectedESLint);
 
         const expectedTypeScript = [repositories[1]];
-        const actualTypeScript = autoApprover['getMatchingRepositories'](repositories, /typescript-4.0.3/gi);
+        const actualTypeScript = autoMerge['getMatchingRepositories'](repositories, /typescript-4.0.3/gi);
         expect(actualTypeScript).toEqual(expectedTypeScript);
 
         const expectedAll = repositories;
-        const actualAll = autoApprover['getMatchingRepositories'](repositories, /typescript/gi);
+        const actualAll = autoMerge['getMatchingRepositories'](repositories, /typescript/gi);
         expect(actualAll).toEqual(expectedAll);
 
         const expectedNone: Repository[] = [];
-        const actualNone = autoApprover['getMatchingRepositories'](repositories, /no-match/gi);
+        const actualNone = autoMerge['getMatchingRepositories'](repositories, /no-match/gi);
         expect(actualNone).toEqual(expectedNone);
       });
     });
 
-    describe('commentByMatch', () => {
-      test('comments on matching PRs', async () => {
-        vi.spyOn<any, any>(autoApprover, 'postComment');
-        const comment = '@dependabot squash and merge';
+    describe('mergeByMatch', () => {
+      test('merge matching PRs', async () => {
+        vi.spyOn<any, any>(autoMerge, 'putMerge');
 
-        await autoApprover.commentByMatch(/eslint/gi, comment, repositories);
+        await autoMerge.mergeByMatch(/eslint/gi, repositories);
         const expectedRepositoryESLint = repositories[0];
-        expect(autoApprover['postComment']).toHaveBeenCalledWith(
+        expect(autoMerge['putMerge']).toHaveBeenCalledWith(
           expectedRepositoryESLint.repositorySlug,
           expectedRepositoryESLint.pullRequests[0].number,
-          comment
+          false
         );
 
-        await autoApprover.commentByMatch(/typescript-4.0.3/gi, comment, repositories);
+        await autoMerge.mergeByMatch(/typescript-4.0.3/gi, repositories);
         const expectedRepositoryTypescript = repositories[1];
-        expect(autoApprover['postComment']).toHaveBeenCalledWith(
+        expect(autoMerge['putMerge']).toHaveBeenCalledWith(
           expectedRepositoryTypescript.repositorySlug,
           expectedRepositoryTypescript.pullRequests[0].number,
-          comment
+          false
         );
       });
     });