+++ packages/auto-merge/src/types/index.d.ts
@@ -1,0 +1,5 @@
+export * from './ActionResult.js';
+export * from './AutoMergeConfig.js';
+export * from './GitHubPullRequest.js';
+export * from './Repository.js';
+export * from './RepositoryResult.js';