+++ packages/electron-info/src/cli.ts
@@ -2,7 +2,6 @@
 
 import fs from 'node:fs';
 import path from 'node:path';
-import {fileURLToPath} from 'node:url';
 import {program as commander} from 'commander';
 
 import {ElectronInfo, RawDeps, SupportedDependencies} from './ElectronInfo.js';
@@ -13,8 +12,7 @@ interface PackageJson {
   version: string;
 }
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+const __dirname = import.meta.dirname;
 const packageJsonPath = path.join(__dirname, '../package.json');
 
 const {description, name, version}: PackageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf-8'));