+++ packages/my-timezone/src/cli.ts
@@ -2,25 +2,23 @@
 
 import fs from 'node:fs';
 import path from 'node:path';
-import {fileURLToPath} from 'node:url';
 import {program as commander} from 'commander';
 
 import {MyTimezone} from './MyTimezone.js';
 
 interface PackageJson {
-  bin: Record<string, string>;
   description: string;
+  name: string;
   version: string;
 }
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+const __dirname = import.meta.dirname;
 const packageJsonPath = path.join(__dirname, '../package.json');
 
-const {bin, description, version}: PackageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf-8'));
+const {description, name, version}: PackageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf-8'));
 
 commander
-  .name(Object.keys(bin)[0])
+  .name(name)
   .version(version)
   .description(`${description}\nUse a city name or longitude value as location.`)
   .option('-o, --offline', 'Work offline (default is false)')