+++ packages/publish-flat/src/cli-copy.ts
@@ -2,7 +2,6 @@
 
 import fs from 'node:fs';
 import path from 'node:path';
-import {fileURLToPath} from 'node:url';
 import {program as commander} from 'commander';
 
 import {copyJson} from './copyJson.js';
@@ -12,8 +11,7 @@ interface PackageJson {
   version: string;
 }
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+const __dirname = import.meta.dirname;
 const packageJsonPath = path.join(__dirname, '../package.json');
 
 const {bin, version}: PackageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf-8'));